changes in data archival code

+++ Lambda-function-code/lambda_function_data_archive.py
@@ -16,17 +16,18 @@ def lambda_handler(event, context):
         s3 = boto3.client('s3')
         
         for obj in s3.list_objects_v2(Bucket=source_bucket)['Contents']:
-            source_key = obj['Key']
-            destination_key = source_key
-            s3.copy_object(CopySource={'Bucket':source_bucket,'Key':source_key},
+            if 'transactions_' in obj['Key']:
+                source_key = obj['Key']
+                destination_key = source_key
+                s3.copy_object(CopySource={'Bucket':source_bucket,'Key':source_key},
                             Bucket=destination_bucket,Key=destination_key)
-            print(f"Copied {source_key} to {destination_bucket}/{destination_key}")
-            s3.delete_object(Bucket=source_bucket, Key=source_key)
-            print(f"Deleted the Object as well")
-            
+                print(f"Copied {source_key} to {destination_bucket}/{destination_key}")
+                s3.delete_object(Bucket=source_bucket, Key=source_key)
+                print("Deleted the Object as well")
+
     
     
     return {
         'statusCode': 200,
-        'body': json.dumps('Hello from Lambda!')
+        'body': json.dumps('Data Archived Successfully')
     }
